courses/build_production_ready_agents/ch2_lab/requirements.txt
google-cloud-aiplatform
google-adk
fastapi
uvicorn
httpx
asyncpg